face-recognition_function/requirements.txt
boto3
facenet-pytorch
opencv-python-headless
pillow